Change time varying gain tests to use pytest.
Keywords in vary_gains_in_time were made singular, rather than plural.
These changes were reflected in the tests, and the tests for the
vary_gains_in_time function were modified to use pytest. Things are
broken right now; trying to convert the rest of the test module to
pytest to see if that fixes some of the problems (there's very weird
behavior of pytest fixtures that is unexpected, and I have a suspicion
that it's because it doesn't want to play nice with unittest, or
vice-versa).