Convert test_sigchain to pytest-only.
All of the old tests were converted from using unittest/nose to using
only pytest. Issues with the tests that were present in the previous
commit have been resolved; the unusual errors that were mentioned in the
previous commit were due to a failure to provide some test functions with
the fixtures they were using. Some of the original tests have been
cleaned up using new Fourier transform tools from uvtools; however, the
tests could be cleaned up even further by consolidating some of the
fixtures used for the old tests and those used for the time variation
tests. It may be possible to make some of the test module more compact
by exploiting parameterization, but I'll leave that for a future PR.