Remove xtalk symmetry enforcement.
Undo the change that enforced that crosstalk is symmetric in delay. Add
an option in the xtalk spectrum class to symmetrize the spectrum (i.e.
make the spectrum also appear at the negative of the delay range
specified). Add a note regarding differences in the parameter jitter
between the original crosstalk spectrum and the negative delay copy.