Increase tolerance for a variation test.
The test for noiselike variation of gain delays failed due to the
calculated standard deviation of delays not being close enough to the
expected standard deviation. This is likely due to the nonlinear map
from delays to gains messing up the distribution, so I've increased the
tolerance for that particular test to try to account for this.