Add a TODO in eor.noiselike_eor.
In modifying the tests for the eor module, I noticed that the covariance
for the noiselike_eor *autocorrelation* visibilities looked very wrong.
That means that the dirty trick for ensuring real-valued autocorrelation
visibilities for noiselike EoR actually messes up the statistics in a
pretty bad way. So I left a reminder that I need to think a bit harder
about how to properly conjure up autocorrelation visibilities for a
noiselike signal. The problem likely has to do with the real part being
taken after fringe-rate and delay filters have been applied... but it
needs a closer investigation.