Update foreground module tests.
Converted all tests for the foregrounds module to use pytest. The
orientation testing function is a little wonky and is expected to fail
for point source foregrounds (with the current implementation), but I
couldn't figure out how to make pytest.mark.xfail play nicely with
pytest.parametrize when only some of the settings are expected to fail.
There is a comment in the test that notes this.