Add comment about a problem with Tsky interpolator
The Tsky interpolator object currently assumes that any data it is
provided covers all LSTs and the domain is extended to account for phase
wraps in accordance with that assumption. This is, in general, not a
good assumption, and checking should be added to ensure that this is
only performed where appropriate.