Convert test_rfi to pytest-only. Fix bugs in DTV.
The new RFI tests are more robust than the previous tests, and they also
test for some warnings/errors being raised as expected. Updating the
tests revealed some subtle bugs with DTV simulation when the observed
frequency channels perfectly align with the DTV sub-band edges; these
were fixed, but it was kind of a clunky fix.