Update test_utils to pytest-only.
This commit completes the transfer of utility module tests to
pytest-only. Most of the tests have been rewritten in an attempt to make
the purpose of each test more clear. There is an outstanding TODO
regarding one test, but that is for this author's understanding of why
the test works.