fix: make Simulator.get method actually work
This required a few changes to various parts of the code in order to
work. The parameter recording in the add method was updated so that the
correct parameters could be retrieved in the get method. During the
process, some improvements/changes elsewhere were also made. The
_generate_redundant_seed method was removed as it was... redundant. The
parameter initialization methods were updated to only pull positional
arguments (rather than all parameters). Various properties were made
into cached properties, and the _get_reds method was converted to a
cached property ``reds``. A few additional changes were made in
anticipation of adding full polarization support.