update: ensure all simulated effects recoverable
This update does some playing with the random seeds so that a previously
simulated effect may be recovered using the ``get`` method even when a
random seed was not provided and the effect generates random numbers.
This commit also sneaks in an update that allows for simulating an
effect multiple times with different parameters to have all of those
parameters tracked (the components can be "named" when added with the
``add`` method).