update: components don't *need* a class doc
Prior to this update, trying to make a component from a class without a
docstring would error out. Even though it isn't great practice to write
a class without a docstring, users shouldn't be *required* to write a
docstring just to use the component feature, so this commit fixes that.