fix: don't convert baseline units in pntsrc_fg
The baseline vector passed to the PointSourceForeground's call method
was being converted to nanoseconds prior to this update. This behavior
was inconsistent with the documentation, as well as other classes, which
use the convention that the baseline vector is already provided in units
of nanoseconds. This commit fixes that discrepancy.