fix: inferred integration time in ThermalNoise
Without this fix, the inferred integration time for a phase-wrapped LST
array is negative for most use cases, which causes the calculated noise
to take on nan values. This commit fixes that issue, but it also adds an
exception for the case where the LSTs are phase-wrapped and the sky
temperature is intended to be calculated from an interpolation object.
(This would have been handled poorly prior to this fix, so this just
makes the error more explicit.)