fix: coupling coefficient calculation
The Jones matrix reordering step was fixed so that the shapes used in
calculation are (n_directions, n_freqs, n_feeds, n_axes). Prior to this,
the last two axes were swapped.

The overall scaling of the coupling coefficients was also not updated
correctly when the calculation was updated a few commits ago--it was
scaled by the frequency rather than the wavelength. This has been fixed.