fix: jit reshape issues
Apparently, when numba compiles the code, any expression a[...] = b[...]
is compiled assuming that a and b have the same number of dimensions.
It's unclear whether this was the case in earlier versions of numba, but
if it was, then it was a lucky coincidence that the old data array
shapes had the same number of dimensions as the visibility matrices used
in calculating the mutual coupling.

Anyway, this is fixed, and so the reshaping tests are fixed. One more
fix still needs to be made.