update: Simulator logic for handling random states
The basic idea here is that the random state seeding is handled by the
Simulator, then the random number generator produced by the state
handling code is passed off to the SimulationComponent to be used in
simulating data. The changes added in this commit will not work until
the SimulationComponent children classes are updated to handle this
difference in how random number generation is handled.